exposed streaming
